Crate colors

rcolours

CSS and XKCD colour package for rust.

Makes CSS and XKCD colours (copied shamelessly from matplotlib's color module) available as u8-tuples and provides basic conversion functions into different representations.

An emphasis is on transforming the colour tuple to ANSI escape sequences for colouring console output. For this goal the crate provides the

Colour representation

For each colour there is detailled information on how to represent it in different ways (for now hex, int/float tuple in RGB space). Other colour spaces (HUV, L*AB, maybe also XYZ and spectrum samples) to come).

Basic Usage

use colors::clr::RED;
use colors::Ansi;

println!("Change colour and font face of console output:");
println!("Make the output {}red{}, {}bold{}", Ansi::ColorEscape(RED), Ansi::Reset, Ansi::Bold, Ansi::Reset);

Modules

clr
Module containing all the colour definitions (and the main struct Color, of course).

Enums

Ansi
Different kinds of escape codes. The ColorEscape variant converts a Color struct into an escape code.

Traits

AnsiEscape
A trait that transforms an objects into an ANSI escape sequence.